refactor: Move generic chart helper and types into utils
