+++ src/app.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useColorMode } from '@chakra-ui/react';
 
 import { ProductTokensBySymbol } from './constants/tokens';
 
@@ -7,14 +6,6 @@ import TokenAllocations from './widgets/token-allocations/TokenAllocations';
 import TokenLineCharts from './widgets/token-line-charts/TokenLineCharts';
 
 const App = ({ domElement }: { domElement: Element }) => {
-  // Local theme helper (light/dark)
-  if (process.env.NODE_ENV === 'development') {
-    const { colorMode, toggleColorMode } = useColorMode();
-    console.debug(`App ~ colorMode ${colorMode}. toggleColorMode() to toggle`);
-    // @ts-ignore
-    window.toggleColorMode = toggleColorMode;
-  }
-
   const widgetType = domElement.getAttribute('data-widget-type');
   const tokenSymbol = domElement.getAttribute(
     'data-token-symbol'
