+++ src/utils/chart.ts
@@ -1,0 +1,22 @@
+export interface ChartOption {
+  width?: number;
+  height?: number;
+  hideYAxis?: boolean;
+  lineColor?: string;
+}
+
+export type ChartDatas = ChartDataPoint[];
+
+export interface ChartDataPoint {
+  x: number;
+  y: number;
+  y2?: number;
+  y3?: number;
+  y4?: number;
+  y5?: number;
+}
+
+/**
+ * Max points for a line chart before degradation in performance
+ */
+export const MaxPointsPerRange: number = 200;
