+++ src/widgets/token-line-charts/token-price/TokenPrice.tsx
@@ -5,6 +5,7 @@ import {
   TokenMarketDataValues,
   selectLatestMarketData,
 } from '../../../providers/MarketData';
+import { ChartOption } from '../../../utils/chart';
 
 import { MaxPanelWidth } from '../TokenLineCharts';
 import TokenPriceChart from './TokenPriceChart';
@@ -28,24 +29,12 @@ export enum PriceChartRangeOption {
   QUARTERLY_PRICE_RANGE = 90,
 }
 
-export interface MarketChartOptions {
-  width?: number;
-  height?: number;
-  hideYAxis?: boolean;
-  lineColor?: string;
-}
-
-export interface PriceChartData {
-  x: number;
-  y: number;
-}
-
 const TokenPrice = ({
   marketData,
   options,
 }: {
   marketData: TokenMarketDataValues;
-  options?: MarketChartOptions;
+  options?: ChartOption;
 }) => {
   const priceChartData = getPriceChartData([marketData]);
